move k8sprocessor -> k8sattributesprocessor
