+++ processor/k8sattributesprocessor/README.md
@@ -1,0 +1,1 @@
+Documentation is published to [pkg.go.dev](https://pkg.go.dev/github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor?tab=doc)
