+++ processor/k8sattributesprocessor/kube/client.go
@@ -31,7 +31,7 @@ import (
 	"k8s.io/client-go/tools/cache"
 
 	"github.com/open-telemetry/opentelemetry-collector-contrib/internal/k8sconfig"
-	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sprocessor/observability"
+	"github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sattributesprocessor/observability"
 )
 
 // WatchClient is the main interface provided by this package to a kubernetes cluster.
