+++ processor/k8sattributesprocessor/testdata/config.yaml
