--- processor/k8sprocessor/README.md
@@ -1,1 +1,0 @@
-Documentation is published to [pkg.go.dev](https://pkg.go.dev/github.com/open-telemetry/opentelemetry-collector-contrib/processor/k8sprocessor?tab=doc)
